refactor(step2): 분산된 로또 번호 생성 책임 병합 및 객체 생성 방식 제한
